# openapi.yaml
...
openapi: 3.0.1
info:
  title: IDS AppStore
  description: IDS AppStore developed by the Fraunhofer FIT
  contact:
    name: IDS AppStore
    url: https://spring.io/projects/spring-boot/appstore
    email: '@email@'
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
  version: 1.0.0
servers:
  - url: https://localhost:8080
    description: Generated server url
...
paths:
...
  /api/registry/user:
    get:
      tags:
        - "Registry: User and Container management"
      summary: Create user credentials for pulling images
      operationId: createUserCredentialsForPullingImages
      responses:
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: object
        "201":
          description: Created
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
        - "Registry: User and Container management"
      summary: Create new user
      description: "Creates a new user in the registry. Password or secret must be\
        \ longer than 8 chars with at least 1 uppercase letter, 1 lowercase letter\
        \ and 1 number"
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationReq'
        required: true
      responses:
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: object
        "201":
          description: Created
          content:
            '*/*':
              schema:
                type: object
  /api/registry/project/member:
    post:
      tags:
        - "Registry: User and Container management"
      summary: Create project member
      operationId: createProjectMember
      parameters:
        - name: projectUserRole
          in: query
          description: "The project members role. The role id 1 for projectAdmin, 2\
          \ for developer, 3 for guest, 4 for maintainer."
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEntity'
        required: true
      responses:
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                type: object
...
  /api/registry/users:
    get:
      tags:
        - "Registry: User and Container management"
      summary: Get user list
      description: Gets the registry user list
      operationId: getRegistryUsers
      responses:
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                type: object
  /api/connector:
    get:
      tags:
        - Connector
      summary: Private IDS self-description
      operationId: getPrivateSelfDescription
      responses:
        "401":
          description: Unauthorized
          content:
            application/ld+json:
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            application/ld+json:
              schema:
                type: object
        "200":
          description: Ok
          content:
            application/ld+json:
              schema:
                type: object
...
  /:
    get:
      tags:
        - Connector
      summary: Public IDS self-description
      operationId: getPublicSelfDescription_1
      responses:
        "401":
          description: Unauthorized
          content:
            application/ld+json:
              schema:
                type: object
        "200":
          description: Ok
          content:
            application/ld+json:
              schema:
                type: object
components:
  schemas:
...
    UserCreationReq:
      type: object
      properties:
        email:
          type: string
        realname:
          type: string
        comment:
          type: string
        password:
          type: string
        username:
          type: string
      description: The user
    UserEntity:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        username:
          type: string
      description: The project member
...